+++ src/package/vendors/pdf/index.tsx
@@ -4,7 +4,14 @@ import type { FileRenderContext } from '@/package/common/type'
 const PdfView = defineAsyncComponent(() => import('./PdfView.vue'))
 export default async function renderPdf(buffer: ArrayBuffer, target: HTMLDivElement, context?: FileRenderContext) {
   const app = createApp({
-    render: () => <PdfView data={buffer} exportAdapter={context?.registerExportAdapter} options={context?.options?.pdf} />
+    render: () => (
+      <PdfView
+        data={buffer.byteLength ? buffer : undefined}
+        url={context?.streamUrl}
+        exportAdapter={context?.registerExportAdapter}
+        options={context?.options?.pdf}
+      />
+    )
   })
   app.mount(target)
   return app